+++ babel.config.js
@@ -1,4 +1,4 @@
 module.exports = {
-  presets: [["@babel/env"]],
+  presets: ["@babel/env"],
   plugins: [["polyfill-corejs3", { method: "usage-pure" }]],
 };
