+++ tsconfig.json
@@ -7,7 +7,8 @@
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true
+    "forceConsistentCasingInFileNames": true,
+    "newLine": "lf" // for rollup-plugin-typescript2
   },
   // see also: https://github.com/ezolenko/rollup-plugin-typescript2/issues/211
   "include": ["src/**/*"],
